feat(graphin):clean code and update layout controller
